Refactor model config test cases
Refactor model config tests as per functionality
